refactor(session): compress restart recovery internals
