+++ src/agents/main-session-recovery-state.ts
@@ -355,27 +355,7 @@ export function transitionMainSessionRecovery(
         },
       };
     }
-    case "cancel_reservation": {
-      const state = entry.mainRestartRecovery;
-      const reserved = state?.reservation;
-      if (
-        !state ||
-        entry.sessionId !== command.reservation.sessionId ||
-        state.cycleId !== command.reservation.cycleId ||
-        reserved?.runId !== command.reservation.runId ||
-        reserved.attempt !== command.reservation.attempt ||
-        reserved.lifecycleGeneration !== command.reservation.lifecycleGeneration
-      ) {
-        return { kind: "rejected", reason: "stale_reservation" };
-      }
-      entry.mainRestartRecovery = {
-        ...state,
-        revision: nextRevision(state),
-        chargedAttempts: Math.max(0, command.reservation.attempt - 1),
-        reservation: undefined,
-      };
-      return { kind: "applied" };
-    }
+    case "cancel_reservation":
     case "abandon_reservation": {
       const state = entry.mainRestartRecovery;
       const reserved = state?.reservation;
@@ -392,6 +372,10 @@ export function transitionMainSessionRecovery(
       entry.mainRestartRecovery = {
         ...state,
         revision: nextRevision(state),
+        chargedAttempts:
+          command.kind === "cancel_reservation"
+            ? Math.max(0, command.reservation.attempt - 1)
+            : state.chargedAttempts,
         reservation: undefined,
       };
       return { kind: "applied" };
