+++ src/agents/main-session-restart-recovery-failure.ts
@@ -104,16 +104,14 @@ export async function tombstoneMainRestartRecoveryWithNotice(params: {
   if (!tombstonedEntry) {
     return "skipped";
   }
-  if (deliveryContext) {
-    await sendUnresumableSessionNotice({
-      deliveryContext,
-      entry: tombstonedEntry,
-      gatewayRuntime: params.gatewayRuntime,
-      reason: params.reason,
-      sessionKey: params.sessionKey,
-      text: TOMBSTONED_SESSION_NOTICE,
-    });
-  }
+  await sendUnresumableSessionNotice({
+    deliveryContext,
+    entry: tombstonedEntry,
+    gatewayRuntime: params.gatewayRuntime,
+    reason: params.reason,
+    sessionKey: params.sessionKey,
+    text: TOMBSTONED_SESSION_NOTICE,
+  });
   return "tombstoned";
 }
