+++ src/agents/main-session-restart-recovery.ts
@@ -36,12 +36,7 @@ import {
 } from "../infra/agent-events.js";
 import { createSubsystemLogger } from "../logging/subsystem.js";
 import { runWithGatewayIndependentRootWorkAdmission } from "../process/gateway-work-admission.js";
-import {
-  isAcpSessionKey,
-  isCronSessionKey,
-  isSubagentSessionKey,
-  resolveAgentIdFromSessionKey,
-} from "../routing/session-key.js";
+import { resolveAgentIdFromSessionKey } from "../routing/session-key.js";
 import {
   beginSessionWorkAdmission,
   cancelSessionWorkAdmissionHandoff,
@@ -112,15 +107,7 @@ function loadExpectedRestartRecoveryTarget(params: {
 }
 
 function shouldSkipMainRecovery(entry: SessionEntry, sessionKey: string): boolean {
-  if (typeof entry.spawnDepth === "number" && entry.spawnDepth > 0) {
-    return true;
-  }
-  if (entry.subagentRole != null) {
-    return true;
-  }
-  return (
-    isSubagentSessionKey(sessionKey) || isCronSessionKey(sessionKey) || isAcpSessionKey(sessionKey)
-  );
+  return !isMainRestartRecoveryCandidate(entry, sessionKey);
 }
 
 function normalizeStringSet(values: Iterable<string> | undefined): Set<string> {
@@ -1391,7 +1378,6 @@ async function writeUnresumableSessionNotice(params: {
 async function failUnresumableMainSession(params: {
   cfg?: OpenClawConfig;
   entry: SessionEntry;
-  expectedRecoverySourceRunId?: string;
   gatewayRuntime: GatewayRecoveryRuntime;
   observation: MainSessionRecoveryObservation;
   reason: string;
@@ -1683,9 +1669,6 @@ async function recoverStore(params: {
       const disposition = await failUnresumableMainSession({
         cfg: params.cfg,
         entry,
-        expectedRecoverySourceRunId: normalizeOptionalString(
-          entry.restartRecoveryDeliverySourceRunId,
-        ),
         gatewayRuntime: params.gatewayRuntime,
         observation: recoveryView.observation,
         reason: "message-tool-only recovery authority is unavailable",
@@ -1716,7 +1699,6 @@ async function recoverStore(params: {
       const disposition = await failUnresumableMainSession({
         cfg: params.cfg,
         entry,
-        expectedRecoverySourceRunId,
         gatewayRuntime: params.gatewayRuntime,
         observation: recoveryView.observation,
         reason: resumeBlockReason,
@@ -1848,7 +1830,6 @@ async function recoverStore(params: {
         const disposition = await failUnresumableMainSession({
           cfg: params.cfg,
           entry,
-          expectedRecoverySourceRunId,
           gatewayRuntime: params.gatewayRuntime,
           observation: recoveryView.observation,
           reason: completion.reason,
@@ -1863,7 +1844,6 @@ async function recoverStore(params: {
       const disposition = await failUnresumableMainSession({
         cfg: params.cfg,
         entry,
-        expectedRecoverySourceRunId,
         gatewayRuntime: params.gatewayRuntime,
         observation: recoveryView.observation,
         reason: resumePolicy.reason,
