+++ src/plugin-sdk/session-store-runtime.ts
@@ -17,9 +17,11 @@ import {
   loadTranscriptEventsSync as loadAccessorTranscriptEventsSync,
   listSessionEntries as listAccessorSessionEntries,
   loadSessionEntry,
+  patchSessionEntry as patchAccessorSessionEntry,
   readSessionUpdatedAt as readAccessorSessionUpdatedAt,
   readTranscriptStatsSync as readAccessorTranscriptStatsSync,
   resolveTranscriptSessionKeyBySessionId as resolveAccessorTranscriptSessionKeyBySessionId,
+  updateSessionEntry,
 } from "../config/sessions/session-accessor.js";
 import { resolveSqliteTargetFromSessionStorePath } from "../config/sessions/session-sqlite-target.js";
 import {
@@ -27,6 +29,7 @@ import {
   parseSqliteSessionFileMarker,
 } from "../config/sessions/sqlite-marker.js";
 import { resolveSessionStoreEntry as resolveSessionStoreEntryFromStore } from "../config/sessions/store-entry.js";
+import { normalizeResolvedMaintenanceConfigInput } from "../config/sessions/store-maintenance.js";
 import type { ResolvedSessionMaintenanceConfigInput } from "../config/sessions/store.js";
 import type {
   AmbientTranscriptWatermark,
@@ -36,17 +39,15 @@ import type {
 import { replaceFileAtomicSync } from "../infra/replace-file.js";
 import { resolveAgentIdFromSessionKey } from "../routing/session-key.js";
 import {
+  activeRecoveryFieldsForSameSession,
+  clearRecoveryStateForRotatedSessionPatch,
   projectPluginSessionEntry,
+  projectPluginSessionEntryPatch,
   projectPluginSessionStore,
   reconcilePluginSessionStore,
   type SessionStoreReadParams,
   toSessionAccessScope,
 } from "./session-store-runtime-internal.js";
-import {
-  patchPluginSessionEntry,
-  updatePluginSessionStoreEntry,
-  upsertPluginSessionEntry,
-} from "./session-store-runtime-mutations.js";
 import type { SessionTranscriptEvent } from "./session-transcript-runtime.js";
 
 const SQLITE_SESSION_STORE_BACKUP_SUFFIXES = ["", "-wal", "-shm", "-journal"] as const;
@@ -130,6 +131,19 @@ type SessionLifecycleArtifactsCleanupResult = {
   archivedTranscriptArtifacts: number;
   removedEntries: number;
 };
+
+function preserveCoreRecoveryState(
+  persistedEntry: InternalSessionEntry,
+  publicPatch: Partial<SessionEntry>,
+): Partial<InternalSessionEntry> {
+  const nextSessionId = Object.hasOwn(publicPatch, "sessionId")
+    ? publicPatch.sessionId
+    : persistedEntry.sessionId;
+  const recoveryState = activeRecoveryFieldsForSameSession(persistedEntry, nextSessionId);
+  return recoveryState
+    ? { ...publicPatch, ...recoveryState }
+    : clearRecoveryStateForRotatedSessionPatch(persistedEntry, publicPatch);
+}
 
 function resolveLegacySessionStoreTarget(storePath: string): {
   agentId?: string;
@@ -397,7 +411,35 @@ export function resolveTranscriptSessionKeyBySessionId(params: {
 export async function patchSessionEntry(
   params: PatchSessionEntryParams,
 ): Promise<SessionEntry | null> {
-  return await patchPluginSessionEntry(params);
+  const entry = await patchAccessorSessionEntry(
+    toSessionAccessScope(params),
+    async (internalEntry, context) => {
+      const persistedEntry = internalEntry as InternalSessionEntry;
+      const patch = await params.update(projectPluginSessionEntry(internalEntry), {
+        existingEntry: context.existingEntry
+          ? projectPluginSessionEntry(context.existingEntry)
+          : undefined,
+      });
+      if (!patch) {
+        return null;
+      }
+      return preserveCoreRecoveryState(persistedEntry, projectPluginSessionEntryPatch(patch));
+    },
+    {
+      fallbackEntry: params.fallbackEntry
+        ? projectPluginSessionEntry(params.fallbackEntry)
+        : undefined,
+      maintenanceConfig:
+        params.maintenanceConfig !== undefined
+          ? normalizeResolvedMaintenanceConfigInput(params.maintenanceConfig)
+          : undefined,
+      preserveActivity: params.preserveActivity,
+      requireWriteSuccess: params.requireWriteSuccess,
+      replaceEntry: params.replaceEntry,
+      skipMaintenance: params.skipMaintenance,
+    },
+  );
+  return entry ? projectPluginSessionEntry(entry) : null;
 }
 
 /** Reads the last activity timestamp for one session entry. */
@@ -418,12 +460,36 @@ export function readAmbientTranscriptWatermark(
 export async function updateSessionStoreEntry(
   params: UpdateSessionStoreEntryParams,
 ): Promise<SessionEntry | null> {
-  return await updatePluginSessionStoreEntry(params);
+  const entry = await updateSessionEntry(
+    { sessionKey: params.sessionKey, storePath: params.storePath },
+    async (internalEntry) => {
+      const patch = await params.update(projectPluginSessionEntry(internalEntry));
+      if (!patch) {
+        return null;
+      }
+      const persistedEntry = internalEntry as InternalSessionEntry;
+      return preserveCoreRecoveryState(persistedEntry, projectPluginSessionEntryPatch(patch));
+    },
+    {
+      skipMaintenance: params.skipMaintenance,
+      takeCacheOwnership: params.takeCacheOwnership,
+      requireWriteSuccess: params.requireWriteSuccess,
+    },
+  );
+  return entry ? projectPluginSessionEntry(entry) : null;
 }
 
 /** Replaces or creates one session entry by agent/session identity. */
 export async function upsertSessionEntry(params: UpsertSessionEntryParams): Promise<void> {
-  await upsertPluginSessionEntry(params);
+  const publicEntry = projectPluginSessionEntry(params.entry);
+  await patchAccessorSessionEntry(
+    toSessionAccessScope(params),
+    (internalEntry) => {
+      const persistedEntry = internalEntry as InternalSessionEntry;
+      return preserveCoreRecoveryState(persistedEntry, publicEntry);
+    },
+    { fallbackEntry: publicEntry, replaceEntry: true },
+  );
 }
 
 /** Deletes one session entry by agent/session identity. */
